decouple themes and extension abstractions from hot reload logic
