+++ lib/shopify_cli/theme/dev_server/hot_reload.rb
@@ -1,27 +1,18 @@
 # frozen_string_literal: true
 
-require_relative "hot_reload/remote_file_reloader"
-require_relative "hot_reload/remote_file_deleter"
-require_relative "hot_reload/sections_index"
-
 module ShopifyCLI
   module Theme
     class DevServer
       class HotReload
-        def initialize(ctx, app, theme:, watcher:, mode:, ignore_filter: nil, extension: nil)
+        def initialize(ctx, app, broadcast_hooks: [], js_hook: nil, watcher:, mode:)
           @ctx = ctx
           @app = app
-          @theme = theme
           @mode = mode
+          @broadcast_hooks = broadcast_hooks
+          @js_hook = js_hook
           @streams = SSE::Streams.new
-          @remote_file_reloader = RemoteFileReloader.new(ctx, theme: @theme, streams: @streams)
-          @remote_file_deleter = RemoteFileDeleter.new(ctx, theme: @theme, streams: @streams)
-          @sections_index = SectionsIndex.new(@theme)
           @watcher = watcher
           @watcher.add_observer(self, :notify_streams_of_file_change)
-          @ignore_filter = ignore_filter
-          @extension = extension
-          @extension_mode = !extension.nil?
         end
 
         def call(env)
@@ -41,80 +32,19 @@ module ShopifyCLI
         end
 
         def notify_streams_of_file_change(modified, added, removed)
-          files = (modified + added)
-            .reject { |file| @ignore_filter&.ignore?(file.relative_path) }
-            .map { |file| @extension_mode ? @extension[file] : @theme[file] }
-
-          files -= liquid_css_files = files.select(&:liquid_css?)
-
-          deleted_files = removed
-            .map { |file| @theme[file] }
-            .reject { |file| @ignore_filter&.ignore?(file.relative_path) }
-
-          remote_delete(deleted_files) unless deleted_files.empty?
-          reload_page(removed) unless deleted_files.empty?
-
-          hot_reload(files) unless files.empty?
-          remote_reload(liquid_css_files)
+          @broadcast_hooks.each do |hook|
+            hook.call(modified, added, removed, streams: @streams)
+          end
         end
 
         private
-
-        def hot_reload(files)
-          paths = files.map(&:relative_path)
-          @streams.broadcast(JSON.generate(modified: paths))
-          @ctx.debug("[HotReload] Modified #{paths.join(", ")}")
-        end
-
-        def reload_page(removed)
-          @streams.broadcast(JSON.generate(reload_page: true))
-          @ctx.debug("[ReloadPage] Deleted #{removed.join(", ")}")
-        end
-
-        def remote_delete(files)
-          files.each do |file|
-            @ctx.debug("delete file each -> file.relative_path #{file.relative_path}")
-            @remote_file_deleter.delete(file)
-          end
-        end
-
-        def remote_reload(files)
-          files.each do |file|
-            @ctx.debug("reload file each -> file.relative_path #{file.relative_path}")
-            @remote_file_reloader.reload(file)
-          end
-        end
 
         def request_is_html?(headers)
           headers["content-type"]&.start_with?("text/html")
         end
 
         def inject_hot_reload_javascript(body)
-          hot_reload_filename = @extension_mode ? "hot-reload-tae.js" : "hot-reload-theme.js"
-          hot_reload_no_script = ::File.read("#{__dir__}/hot-reload-no-script.html")
-          hot_reload_js = ::File.read("#{__dir__}/#{hot_reload_filename}")
-          hot_reload_script = [
-            hot_reload_no_script,
-            "<script>",
-            params_js,
-            hot_reload_js,
-            "</script>",
-          ].join("\n")
-
-          body = body.join.gsub("</body>", "#{hot_reload_script}\n</body>")
-
-          [body]
-        end
-
-        def params_js
-          env = { mode: @mode }
-          env[:section_names_by_type] = @extension_mode ? [] : @sections_index.section_names_by_type
-
-          <<~JS
-            (() => {
-              window.__SHOPIFY_CLI_ENV__ = #{env.to_json};
-            })();
-          JS
+          @js_hook&.call(body: body, dir: __dir__, mode: @mode)
         end
 
         def create_stream
