+++ test/shopify-cli/theme/dev_server/hot_reload_test.rb
@@ -1,9 +1,7 @@
 # frozen_string_literal: true
 require "test_helper"
-require "shopify_cli/theme/dev_server"
-require "shopify_cli/theme/dev_server/hot_reload/remote_file_reloader"
-require "shopify_cli/theme/dev_server/hot_reload/remote_file_deleter"
 require "rack/mock"
+require "shopify_cli/theme/dev_server"
 
 module ShopifyCLI
   module Theme
@@ -21,174 +19,42 @@ module ShopifyCLI
           @mode = "off"
         end
 
-        def test_hot_reload_js_injected_if_html_request
-          html = <<~HTML
-            <html>
-              <head></head>
-              <body>
-                <h1>Hello</h1>
-              </body>
-            </html>
-          HTML
-
-          params_js = <<~JS
-            (() => {
-              window.__SHOPIFY_CLI_ENV__ = {"mode":"off","section_names_by_type":{"main-blog":["main"]}};
-            })();
-          JS
-
-          reload_js = ::File.read(
-            ::File.expand_path("lib/shopify_cli/theme/dev_server/hot-reload-theme.js", ShopifyCLI::ROOT)
-          )
-          hot_reload_no_script = ::File.read(
-            ::File.expand_path("lib/shopify_cli/theme/dev_server/hot-reload-no-script.html", ShopifyCLI::ROOT)
-          )
-
-          injected_script = "<script>\n#{params_js}\n#{reload_js}\n</script>"
-
-          expected_html = <<~HTML
-            <html>
-              <head></head>
-              <body>
-                <h1>Hello</h1>
-              #{hot_reload_no_script}
-            #{injected_script}
-            </body>
-            </html>
-          HTML
-
-          response = serve(html, headers: { "content-type" => "text/html" })
-
-          assert_equal(expected_html, response)
-        end
-
-        def test_does_not_inject_hot_reload_js_for_non_html_responses
-          css = <<~CSS
-            .body { color: red }
-          CSS
-
-          response = serve(css, headers: { "content-type" => "text/css" })
-
-          assert_equal(css, response)
-        end
-
-        def test_streams_on_hot_reload_path
-          SSE::Stream.any_instance.expects(:each).yields("")
-          serve(path: "/hot-reload")
-        end
-
-        def test_broadcasts_watcher_events_when_file_modified
-          modified = ["style.css"]
-          SSE::Streams.any_instance
-            .expects(:broadcast)
-            .with(JSON.generate(modified: modified))
-
+        def test_calls_broadcast_hooks_when_available
           app = -> { [200, {}, []] }
-          HotReload.new(@ctx, app, theme: @theme, watcher: @watcher, mode: @mode)
-
+          @ctx.expects(:debug).with("Ran 1").once
+          @ctx.expects(:debug).with("Ran 2").once
+          @ctx.expects(:debug).with("Ran 3").once
+          HotReload.new(@ctx, app, broadcast_hooks: broadcast_hook_mock, watcher: @watcher, mode: @mode)
           @watcher.changed
-          @watcher.notify_observers(modified, [], [])
+          @watcher.notify_observers([], [], [])
         end
 
-        def test_broadcasts_watcher_events_when_file_deleted
-          deleted = ["announcement.liquid"]
-          HotReload::RemoteFileDeleter
-            .stubs(:new)
-            .returns(remote_file_deleter)
+        def test_calls_reload_js_hook
+          app = lambda do |_env|
+            [200, { "content-type" => "text/html" }, []]
+          end
+          correct_output = "<html><script>console.log('testing');</script></html>"
+          stack = HotReload.new(@ctx, app, watcher: @watcher, mode: @mode, js_hook: reload_js_hook_mock(correct_output))
+          request = Rack::MockRequest.new(stack)
 
-          SSE::Streams.any_instance
-            .expects(:broadcast)
-            .with(JSON.generate(reload_page: true))
-
-          app = -> { [200, {}, []] }
-          HotReload.new(@ctx, app, theme: @theme, watcher: @watcher, mode: @mode)
-
-          @watcher.changed
-          @watcher.notify_observers([], [], deleted)
-        end
-
-        def test_doesnt_broadcast_watcher_events_when_modified_list_is_empty
-          root_path = Pathname.new(__dir__)
-          ignore_filter = ShopifyCLI::Theme::IgnoreFilter.new(root_path, patterns: ["ignored/**"])
-          modified = ["ignored/style.css"]
-          SSE::Streams.any_instance
-            .expects(:broadcast)
-            .with(JSON.generate(modified: modified))
-            .never
-
-          app = -> { [200, {}, []] }
-          HotReload.new(
-            @ctx, app,
-            theme: @theme,
-            watcher: @watcher,
-            mode: @mode,
-            ignore_filter: ignore_filter
-          )
-
-          @watcher.changed
-          @watcher.notify_observers(modified, [], [])
-        end
-
-        def test_doesnt_broadcast_watcher_events_when_deleted_list_is_empty
-          root_path = Pathname.new(__dir__)
-          ignore_filter = ShopifyCLI::Theme::IgnoreFilter.new(root_path, patterns: ["ignored/**"])
-          deleted = ["ignored/announcement.liquid"]
-          SSE::Streams.any_instance
-            .expects(:broadcast)
-            .with(JSON.generate(reload_page: true))
-            .never
-
-          app = -> { [200, {}, []] }
-          HotReload.new(
-            @ctx, app,
-            theme: @theme,
-            watcher: @watcher,
-            mode: @mode,
-            ignore_filter: ignore_filter
-          )
-
-          @watcher.changed
-          @watcher.notify_observers([], [], deleted)
-        end
-
-        def test_doesnt_broadcast_watcher_events_when_modified_file_is_a_liquid_css
-          modified = ["assets/generated.css.liquid"]
-          HotReload::RemoteFileReloader
-            .stubs(:new)
-            .returns(remote_file_reloader)
-          SSE::Streams.any_instance
-            .expects(:broadcast)
-            .with(JSON.generate(modified: modified))
-            .never
-
-          app = -> { [200, {}, []] }
-          HotReload.new(@ctx, app, theme: @theme, watcher: @watcher, mode: @mode)
-
-          @watcher.changed
-          @watcher.notify_observers(modified, [], [])
+          assert_equal(correct_output, request.get("/").body)
         end
 
         private
 
-        def remote_file_reloader
-          reloader = mock("Reloader")
-          reloader.stubs(reload: nil)
-          reloader
+        def app
         end
 
-        def remote_file_deleter
-          deleter = mock("Deleter")
-          deleter.stubs(delete: nil)
-          deleter
+        def reload_js_hook_mock(body)
+          hook = mock("ReloadJSHook", call: body)
+          hook
         end
 
-        def serve(response_body = "", path: "/", headers: {})
-          app = lambda do |_env|
-            [200, headers, [response_body]]
-          end
-          stack = HotReload.new(@ctx, app, theme: @theme, watcher: @watcher, mode: @mode)
-          request = Rack::MockRequest.new(stack)
-          request.get(path).body
+        def broadcast_hook_mock
+          hook1 = mock("BroadcastHook1", call: @ctx.debug("Ran 1"))
+          hook2 = mock("BroadcastHook2", call: @ctx.debug("Ran 2"))
+          hook3 = mock("BroadcastHook3", call: @ctx.debug("Ran 3"))
+          [hook1, hook2, hook3]
         end
       end
     end
